Do not apply a background-color if "null" + allow 0 width
It's impossible to change the background-color via class because the inline color from style is always preferred. Differentiating "undefined" from "null" allow user to specify when they want no background color to be applied.
Also, allowing a 0 width should be possible, for instance to rely instead on borders to display the cursor.